add task model and apis related to task model

+++ src/endpoints/task.py
@@ -23,6 +23,7 @@ class TaskCreate(BaseModel):
     to_date: datetime = Field(..., example="2023-08-15T15:32:00Z")
     from_date: datetime = Field(..., example="2023-08-14T15:32:00Z")
     priority: int = Field(..., example=1)
+    project_key: str = Field(..., example="32ed23f32f2311")
 
 
 class TaskEdit(BaseModel):
@@ -44,20 +45,19 @@ def get_db(request: Request):
     return request.state.db
 
 
-# このインスタンスをアノテーションに利用することでエンドポイントを定義できる
-# app = FastAPI()
-
-
 # taskの全取得
 @router.get("/tasks")
 def get_tasks(db: Session = Depends(get_db)):
     tasks = db.query(Task).all()
+
+    # project_keyでProjectを検索
+
     return tasks
 
 
 # 単一のtaskを取得
 @router.get("/tasks/{task_id}")
-def get_task(task_id: str, db: Session = Depends(get_db)):
+def get_task_by_id(task_id: str, db: Session = Depends(get_db)):
     task = get_task(db, task_id)
     return task
 
@@ -74,6 +74,7 @@ async def create_task(task_created: TaskCreate, db: Session = Depends(get_db)):
         to_date=task_created.to_date,
         from_date=task_created.from_date,
         priority=task_created.priority,
+        project_key=task_created.project_key,
     )
     task.id = str(uuid.uuid4())
     task.created_at = now
@@ -93,6 +94,10 @@ async def update_task(
     now = datetime.now()
 
     task = get_task(db, task_id)
+
+    if not task:
+        return {"error": "Project not found"}, 404
+
     task.title = task_created.title
     task.status = task_created.status
     task.man_hour_min = task_created.man_hour_min
@@ -112,5 +117,9 @@ async def update_task(
 @router.delete("/tasks/{task_id}")
 async def delete_task(task_id: str, db: Session = Depends(get_db)):
     task = get_task(db, task_id)
+
+    if not task:
+        return {"error": "Project not found"}, 404
+
     db.delete(task)
     db.commit()
